backend/pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "LangGraph agent with Anthropic + Replicate"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.4",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "python-dotenv>=1.0",
    "replicate>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
